This is a broken commit
This commit starts working to try and make the Umple plugin behave
similarly to the Scala and Antlr plugins. I've realized that having the
configuration in the `sourceSets` is not worth-while, however, we can use
them to add extra configuration (mostly dependency checks).

The current state has a global configuration closure (`umple`) that
specifies the information on how to generate things _per project_. The
idea would be to allow for the specification of an array of generators and
we always place them into a subdirectory that is
`build/generated-src/$lang`.

Additionally, I've added some testing that does not require the
`integrationTests` style testing, that should simplify some testing -- we
still need the integration ones, though.

[skip ci]