Hopefully one of the last broken commits
[skip-CI]

This commit contains a "proof of concept" that shows how we can store the `UmpleGenerateTask` object as a member variable of the `DefaultUmpleSourceSet`. By maintaining an `UmpleGenerateTask` member variable, we can update the configuration variables stored in `UmpleGenerateTask` once user-defined overrides have been read into the `DefaultUmpleSourceSet`.

This commit also renames any `sourceSet` variables to something else (bb2 and abc123 in `UmpleGradlePlugin.groovy`)

@Nava2, to get this working on your machine, make sure your umple files are located at src/" + sourceSet.getName() + "/umple in your project directory. You should set the `umpleFilePath` in the `umple` closure of the project's `build.gradle` file to point to `master.ump` in the same location.

Here's an example of how to configure the sourceSets in your project's `build.umple` file:

sourceSets {
	   main {
	       umple{
	            umpleFilePath = file('path/to/your/umple/folder')
	       }
	   }

	}
I'm not sure if the path has to be absolute or not. If relative doesn't work, use absolute for now.

Let me know if you have any problems getting things running (you should be able to run things and only get the "No value has been specified for property 'sourceSet'. error").