This is a _mostly working_ commit!
Tried to address some of the concerns from #14, but still needs
integration testing.

@AdamBJ solved most of the leg work, I fought with it a bit more and
figured out how to actually get the information back from the sourceSet
closures.

By adding the properties to the `UmpleSourceSet` (via the `UmpleOptions`
interface), we can now set the properties that we want for running Umple.
This is because when the `Closure` is applied, the `delegate` used for
figuring out what fields are what is set to be the `SourceSet` that it is
called against (in this case, `UmpleSourceSet`). See [1] for more
information on how to apply `Closure`s.

The TODO from all of this:

1. Create a better testing suite -- the "properties-as-strings" approach
is nasty, we need to do better
2. Add more integration testing
3. Verify what I did makes sense.. I think I may have excessively created
the "config-style" as I'm not sure how many tasks are made per project

[1] http://groovy-lang.org/closures.html#closure-this