This is a 90% working commit!
This commit connects a lot of plugin components that weren't communicating properly before. Here's the summary of what I did (see the in-code comments for more information):

Changed the way the UmpleGenerateTask.groovy sets configuration information. Here's what happens now:

The default umple closure (i.e. the umple {} that's declared by the user outside of the sourceSets closure) is created in UmpleGradlePlugin.groovy at line 39. The SourceSet specific version of the umple closure is created for each source set at line 59. If the user has specified custom configuration information it is stored in one of these closures.

Before we go run the UmpleGenerateTask, we first call setUmpleGenerateConfig (formerly loadGlobals) at line 69 of UmpleGenerateTask.groovy. This function checks all the locations where user configuration information could be stored and makes that information available to the UmpleGenerateTask before the task is run. SourceSet specific configuration that the user specifies takes priority over default configuration they specify. If the user hasn't provided SourceSet specific or default configuration information for a particular property we get the configuration information from the hard-coded values in UmpleOptions.groovy.

There are a bunch of other little tweaks included in this commit too, but the configuration change is the big one. You can have a look at the TODO tags scattered throughout the source to see what still needs to be done. I'm going to add some comments on GitHub for the issues I think are more important to resolve.

[skip ci]